terabit_test/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
drf-yasg==1.21.10
psycopg2-binary==2.9.10
python-decouple==3.8
pillow==11.1.0